refactor: removes unused keybinding functionality that is later to be introduced anyway
