--- src/app/producer-shelf/abstractions/keyboard-binding.service.ts
@@ -1,8 +1,0 @@
-import { Subscription } from 'rxjs'
-import { KeyBinding } from '../models/key-binding'
-
-export abstract class KeyboardBindingService {
-  public abstract subscribeToKeybindings(callback: (keyBindings: KeyBinding[]) => void): Subscription
-  public abstract subscribeToPressedKeys(callback: (pressedKeys: string[]) => void): Subscription
-  public abstract unsubscribe(): void
-}
