+++ src/app/producer-shelf/components/producer-shelf/producer-shelf.component.scss
@@ -2,5 +2,4 @@
   position: relative;
   display: block;
   border-top: 2px solid #555;
-  pointer-events: none;
 }
