+++ src/app/producer-shelf/components/producer-shelf/producer-shelf.component.spec.ts
@@ -1,29 +1,14 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing'
 import { ProducerShelfComponent } from './producer-shelf.component'
-import { KeyboardBindingService } from '../../abstractions/keyboard-binding.service'
-import { instance, mock } from '@typestrong/ts-mockito'
-import { KeyboardBindingMatcher } from '../../services/keyboard-binding.matcher'
 
 describe(ProducerShelfComponent.name, () => {
   let component: ProducerShelfComponent
   let fixture: ComponentFixture<ProducerShelfComponent>
 
   beforeEach(async () => {
-    const mockKeyboardBindingService: KeyboardBindingService = mock<KeyboardBindingService>()
-    const mockKeyboardBindingMatcher: KeyboardBindingMatcher = mock<KeyboardBindingMatcher>()
-
     await TestBed.configureTestingModule({
       declarations: [ProducerShelfComponent],
-      providers: [
-        {
-          provide: KeyboardBindingService,
-          useValue: instance(mockKeyboardBindingService),
-        },
-        {
-          provide: KeyboardBindingMatcher,
-          useValue: instance(mockKeyboardBindingMatcher),
-        },
-      ],
+      providers: [],
     }).compileComponents()
 
     fixture = TestBed.createComponent(ProducerShelfComponent)
