--- src/app/producer-shelf/models/key-binding.ts
@@ -1,7 +1,0 @@
-export interface KeyBinding {
-  readonly key: string
-  readonly modifiers: string[]
-  readonly label: string
-  readonly action: () => void
-  readonly onKeyPress: boolean
-}
