+++ src/app/shared/components/draggable-shelf/draggable-shelf.component.html
